% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Data_Input}
\alias{Data_Input}
\title{Input data - MOBASY Building data}
\format{
A data frame of more with more than 700 variables and more than 150 datasets
}
\usage{
Data_Input
}
\description{
Input data of the MOBASY Model
Calculation: MOBASY reality-based model including energy profile estimation procedures
(use of basic monitoring indicators)
TABULA energy performance calculation (physical model),
estimation of uncertainty and target / actual comparison
}
\examples{

# Explanation for a variable can be viewed by use of the data frame "InfoVariables"

myVar <- "d_Insulation_Wall"

paste0 (
  myVar, ": ",
  InfoVariables ["Description_ENG",     myVar], " / ",
  InfoVariables ["Specification_1_ENG", myVar], "; ",
  "unit: ", InfoVariables ["Unit", myVar], "; ",
  "data format: ", InfoVariables ["DataFormat", myVar]
)

# Result:
# "d_Insulation_Wall: Insulation thickness / Walls; unit: cm; data format: Real"

}
\keyword{datasets}
