% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Data_Output}
\alias{Data_Output}
\title{Output data structure}
\format{
A data frame of more with more than 300 variables and more than 150 rows (datasets)
}
\usage{
Data_Output
}
\description{
Empty data frame to receive the results of MOBASY calculation
}
\details{
The output data is assigned to the following sections:

	(8) Model Indicators TABULA Calculation Model

	(8.1) Thermal envelope

	(8.2) Building energy balance (space heating)

	(8.3) Heating system energy balance

	(8.4) DHW energy balance

	(8.5) Summary data energy heating

	(8.6) Summary data energy DHW

	(8.7) Summary data energy

	(8.8) Parameters for the calculation

	(8.9) Comparison Calculation-Metering

	(8.10) Uncertainty of energy balance quantities
}
\examples{

# Explanation for a variable can be viewed by use of the data frame "InfoVariables"

myVar <- "q_Model1_del_sum_gas"

paste0 (
  myVar, ": ",
  InfoVariables ["Description_ENG",     myVar], " / ",
  InfoVariables ["Specification_1_ENG", myVar], "; ",
  "unit: ", InfoVariables ["Unit", myVar], "; ",
  "data format: ", InfoVariables ["DataFormat", myVar]
)

# Result:
# "d_Insulation_Wall: Insulation thickness / Walls; unit: cm; data format: Real"

}
\keyword{datasets}
