% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DF_FilterBuildingData}
\alias{DF_FilterBuildingData}
\title{Predefined filter lists}
\format{
A data frame; the column names are the filter names, the dataset IDs to be used are given in the respective rows
}
\usage{
DF_FilterBuildingData
}
\description{
A dataframe containing filter lists that can be applied to the building data tables
}
\examples{

DF_FilterBuildingData [ , "MOBASY-Sample"]
DF_FilterBuildingData [ , "Example"]
DF_FilterBuildingData [ , "WebTool"]



}
\keyword{datasets}
