% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{InfoVariables}
\alias{InfoVariables}
\title{Information about the variables}
\format{
A data frame of more with more than 1000 variables and 100 rows
}
\usage{
InfoVariables
}
\description{
Building data meta information
}
\examples{

InfoVariables ["Description_ENG",     "A_C_Floor_Intake"]
InfoVariables ["Specification_1_ENG", "A_C_Floor_Intake"]
InfoVariables ["Specification_2_ENG", "A_C_Floor_Intake"]
InfoVariables ["Unit",                "A_C_Floor_Intake"]
InfoVariables ["DataFormat",          "A_C_Floor_Intake"]


# Result:

# "Input floor or footprint area, used for data acquisition"
# "Type of Area defined by selection in \"Code_TypeFloorArea_A_C_Floor_Intake\""
# NA
# "m²"
# "Real"

}
\keyword{datasets}
