% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{myFilterName}
\alias{myFilterName}
\title{The name of the filter that has already been applied}
\format{
A charater string
}
\usage{
myFilterName
}
\description{
A character string specifying the filter that has been applied to create the building data tables given in the package.
The value is usually supposed to be "All" (= no predefined filter applied).
}
\keyword{datasets}
