% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Data_Input}
\alias{Data_Input}
\title{Input data - MOBASY Building data}
\format{
A data frame of more with more than 700 variables and more than 150 datasets
}
\usage{
Data_Input
}
\description{
Input data of the MOBASY Model
Calculation: MOBASY reality-based model including energy profile estimation procedures
(use of basic monitoring indicators)
TABULA energy performance calculation (physical model),
estimation of uncertainty and target / actual comparison
}
\details{
The input data is assigned to the following sections:

	(1) Dataset identification

	(2) General data

	(3) Building Fabric / Basic Monitoring Indicators / Energy Profile Indicators
	(3.1) Basic Geometrical Data
	(3.2) Construction Type
	(3.3) Thermal insulation
	(3.4) Window Type
	(3.5) Thermal Bridging / Airtightness

	(4) Heat Supply System / Basic Monitoring Indicators / Energy Profile Indicators
	(4.1) Central system - heat generation
	(4.2) Central system - heat storage and distribution
	(4.3) Decentral heat generation heating
	(4.4) Decentral heat generation DHW
	(4.5) Further systems (ventilation and PV systems)
	(4.5) Special quality features of heat supply system

	(5) Metered Energy Consumption
	(5.1) Metering Points
	(5.2) Metering periods (consecutive)
	(5.3) Consumption values metering point M1 (main metering)
	(5.4) Consumption values metering point M2 (partial or supplemental metering)
	(5.5) Consumption values metering point M3 (partial or supplemental metering)
	(5.6) Settings for comparison with energy performance calculation

	(6) Data sources / uncertainty of monitoring indidators

	(7) Input TABULA Calculation
	(7.1) Settings / boundary conditions
	(7.2) Alternative manual input for TABULA calculation - thermal envelope
	(7.3) Optional manual input for Energy Profile based calculation - share of produced heat by heat generator type (useful for parameter studies)
	(7.4) Alternative direct input for TABULA calculation - heat supply system
}
\examples{

# Explanation for a variable can be viewed by use of the data frame "InfoVariables"

myVar <- "d_Insulation_Wall"

paste0 (
  myVar, ": ",
  InfoVariables ["Description_ENG",     myVar], " / ",
  InfoVariables ["Specification_1_ENG", myVar], "; ",
  "unit: ", InfoVariables ["Unit", myVar], "; ",
  "data format: ", InfoVariables ["DataFormat", myVar]
)

# Result:
# "d_Insulation_Wall: Insulation thickness / Walls; unit: cm; data format: Real"

}
\keyword{datasets}
