% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Data_Output}
\alias{Data_Output}
\title{Output data structure}
\format{
A data frame of more with more than 300 variables and more than 150 rows (datasets)
}
\usage{
Data_Output
}
\description{
Empty data frame to receive the results of MOBASY calculation
}
\examples{

# Explanation for a variable can be viewed by use of the data frame "InfoVariables"

myVar <- "q_Model1_del_sum_gas"

paste0 (
  myVar, ": ",
  InfoVariables ["Description_ENG",     myVar], " / ",
  InfoVariables ["Specification_1_ENG", myVar], "; ",
  "unit: ", InfoVariables ["Unit", myVar], "; ",
  "data format: ", InfoVariables ["DataFormat", myVar]
)

# Result:
# "d_Insulation_Wall: Insulation thickness / Walls; unit: cm; data format: Real"

}
\keyword{datasets}
